some definitions moved to validation_globals.cpp, added direct includes in validation_... files
